orbtk: cleanup popup_v1 code with popup_v2 code
* orbtk
* orbtk_widgets
